fix additional test tolerances and dtypes

+++ skimage/io/tests/test_imread.py
@@ -28,7 +28,7 @@ def _use_imread_plugin():
 def test_imread_as_gray():
     img = imread(fetch('data/color.png'), as_gray=True)
     assert img.ndim == 2
-    assert img.dtype == np.float64
+    assert img.dtype == np.float32
     img = imread(fetch('data/camera.png'), as_gray=True)
     # check that conversion does not happen for a gray image
     assert np.dtype(img.dtype).char in np.typecodes['AllInteger']
